joke-site-backend/requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.3
sqlalchemy~=2.0.31
asyncpg
pydantic
python-dotenv~=1.0.1
google-auth
google-auth-oauthlib
passlib[bcrypt]
pydantic-settings
starlette~=0.37.2
greenlet~=3.0.3
itsdangerous